+++ src/main/java/com/mrporter/pomangam/admin/controllers/store/image/_StoreImageController.java
@@ -20,8 +20,8 @@ public class _StoreImageController {
 
     @GetMapping
     public ResponseEntity<?> get(
-            @PathVariable(value = "didx", required = true) Integer didx,
-            @PathVariable(value = "sidx", required = true) Integer sidx,
+            @PathVariable(value = "didx", required = true) Long didx,
+            @PathVariable(value = "sidx", required = true) Long sidx,
             @PageableDefault(sort = {"idx"}, direction = Sort.Direction.DESC, size = 10) Pageable pageable
     ) {
         return new ResponseEntity(HttpStatus.OK);
@@ -29,17 +29,17 @@ public class _StoreImageController {
 
     @GetMapping("/{idx}")
     public ResponseEntity<?> getByIdx(
-            @PathVariable(value = "didx", required = true) Integer didx,
-            @PathVariable(value = "sidx", required = true) Integer sidx,
-            @PathVariable(value = "idx", required = true) Integer idx
+            @PathVariable(value = "didx", required = true) Long didx,
+            @PathVariable(value = "sidx", required = true) Long sidx,
+            @PathVariable(value = "idx", required = true) Long idx
     ) {
         return new ResponseEntity(HttpStatus.OK);
     }
 
     @GetMapping("/search/count")
     public ResponseEntity<?> searchCount(
-            @PathVariable(value = "didx", required = true) Integer didx,
-            @PathVariable(value = "sidx", required = true) Integer sidx
+            @PathVariable(value = "didx", required = true) Long didx,
+            @PathVariable(value = "sidx", required = true) Long sidx
     ) {
         return new ResponseEntity(HttpStatus.OK);
     }
