+++ src/main/java/com/mrporter/pomangam/admin/repositories/store/category/_StoreCategoryJpaRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(exported = false)
-public interface _StoreCategoryJpaRepository extends JpaRepository<StoreCategory, Integer> {
+public interface _StoreCategoryJpaRepository extends JpaRepository<StoreCategory, Long> {
 
 }
