+++ src/main/java/com/mrporter/pomangam/admin/repositories/user/_UserJpaRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(exported = false)
-public interface _UserJpaRepository extends JpaRepository<User, Integer> {
+public interface _UserJpaRepository extends JpaRepository<User, Long> {
     User findByPhoneNumber(@Param("phoneNumber") String phoneNumber);
     void deleteByPhoneNumber(@Param("phoneNumber") String phoneNumber);
 }
