+++ src/main/java/com/mrporter/pomangam/client/controllers/deliverysite/detail/DeliveryDetailSiteController.java
@@ -18,22 +18,22 @@ public class DeliveryDetailSiteController {
 
     @GetMapping
     public ResponseEntity<?> get(
-            @PathVariable(value = "didx", required = true) Integer didx
+            @PathVariable(value = "didx", required = true) Long didx
     ) {
         return new ResponseEntity(detailSiteService.getByIdxDeliverySite(didx), HttpStatus.OK);
     }
 
     @GetMapping("/{idx}")
     public ResponseEntity<?> getByIdx(
-            @PathVariable(value = "didx", required = true) Integer didx,
-            @PathVariable(value = "idx", required = true) Integer idx
+            @PathVariable(value = "didx", required = true) Long didx,
+            @PathVariable(value = "idx", required = true) Long idx
     ) {
         return new ResponseEntity(detailSiteService.getByIdx(idx), HttpStatus.OK);
     }
 
     @GetMapping("/search/count")
     public ResponseEntity<?> searchCount(
-            @PathVariable(value = "didx", required = true) Integer didx
+            @PathVariable(value = "didx", required = true) Long didx
     ) {
         return new ResponseEntity(detailSiteService.count(), HttpStatus.OK);
     }
