+++ src/main/java/com/mrporter/pomangam/client/controllers/deliverysite/region/RegionController.java
@@ -21,7 +21,7 @@ public class RegionController {
     }
 
     @GetMapping("/{idx}")
-    public ResponseEntity<?> getByIdx(@PathVariable(value = "idx", required = true) Integer idx
+    public ResponseEntity<?> getByIdx(@PathVariable(value = "idx", required = true) Long idx
     ) {
         return new ResponseEntity(HttpStatus.OK);
     }
