+++ src/main/java/com/mrporter/pomangam/client/controllers/promotion/PromotionController.java
@@ -21,7 +21,7 @@ public class PromotionController {
     }
 
     @GetMapping("/{idx}")
-    public ResponseEntity<?> getByIdx(@PathVariable(value = "idx", required = true) Integer idx
+    public ResponseEntity<?> getByIdx(@PathVariable(value = "idx", required = true) Long idx
     ) {
         return new ResponseEntity(HttpStatus.OK);
     }
