+++ src/main/java/com/mrporter/pomangam/client/repositories/deliverysite/region/RegionJpaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource(exported = false)
-public interface RegionJpaRepository extends JpaRepository<Region, Integer> {
+public interface RegionJpaRepository extends JpaRepository<Region, Long> {
 
 }
